worked on logo component and styletyles
