+++ Busreservation/src/Components/logo/logo.styles.css
@@ -1,0 +1,8 @@
+
+.logo-image-container {
+    display: flex;
+    width: 100px;
+    height: 100px;
+    border-radius: 50%;
+    overflow: hidden;
+}
