+++ Busreservation/src/pages/userTypes/userTypes.styles.css
@@ -1,0 +1,55 @@
+.user-type-container {
+  display: flex;
+  flex-direction: column;
+  width: 100vw;
+  height: 100vh;
+  align-items: center;
+  z-index: 5;
+  background-position: center;
+  background-attachment: fixed;
+}
+
+.logo-container {
+  display: flex;
+  width: 100%;
+  justify-content: center;
+  align-items: center;
+  padding: 1rem;
+  z-index: 1;
+  background: linear-gradient(to bottom right, rgb(26, 235, 250), #fff);
+}
+
+.user-type-button-container {
+  width: 20%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+  padding: 2rem;
+  margin-top: 1rem;
+  z-index: 1;
+}
+
+.user-type-button {
+  width: 100%;
+  padding: 1rem;
+  margin: 0.5rem;
+  border: none;
+  border-radius: 0.5rem;
+  font-size: 1.2rem;
+  font-weight: 600;
+  font: 1rem poppins;
+  cursor: pointer;
+  color: #333;
+}
+
+.background-sheet {
+  display: flex;
+  position: absolute;
+  top: 0;
+  bottom: 0;
+  left: 0;
+  right: 0;
+  background-image: url("../../Assets/images/bus-background.jpg");
+  opacity: 0.5;
+}
